news_etl/requirements.txt
httpx>=0.27
pydantic>=2
azure-ai-textanalytics>=5.3.0
python-dateutil
dateparser
pyyaml
tenacity>=8.2
pandas>=2.0
python-dotenv>=1.0
